+++ src/rendererGLES/tr_scene.c
@@ -589,7 +589,7 @@ void RE_RenderScene(const refdef_t *fd)
 
 	// derived info
 
-	tr.refdef.floatTime = (double)tr.refdef.time * 0.001;
+	tr.refdef.floatTime = tr.refdef.time * 0.001;
 
 	tr.refdef.numDrawSurfs = r_firstSceneDrawSurf;
 	tr.refdef.drawSurfs    = backEndData->drawSurfs;
