+++ src/rendererGLES/tr_shade_calc.c
@@ -34,7 +34,7 @@
 
 #include "tr_local.h"
 
-#define WAVEVALUE(table, base, amplitude, phase, freq)  ((base) + table[(int)((((phase) + tess.shaderTime * (freq)) *FUNCTABLE_SIZE)) & FUNCTABLE_MASK] * (amplitude))
+#define WAVEVALUE(table, base, amplitude, phase, freq)  ((base) + table[(int64_t)((((phase) + tess.shaderTime * (freq)) *FUNCTABLE_SIZE)) & FUNCTABLE_MASK] * (amplitude))
 
 /**
  * @brief TableForFunc
@@ -202,7 +202,7 @@ void RB_CalcDeformVertexes(deformStage_t *ds)
 		}
 		ds->deformationWave.frequency *= -1;
 	}
-	else if (ds->deformationWave.frequency == 0.f)
+	else if (ds->deformationWave.frequency == 0.)
 	{
 		scale = EvalWaveForm(&ds->deformationWave);
 
@@ -247,7 +247,7 @@ void RB_CalcDeformNormals(deformStage_t *ds)
 {
 	int    i;
 	double scale;
-	double *xyz    = ( double * ) tess.xyz;
+	float  *xyz    = ( float * ) tess.xyz;
 	float  *normal = ( float * ) tess.normal;
 
 	for (i = 0; i < tess.numVertexes; i++, xyz += 4, normal += 4)
@@ -281,13 +281,13 @@ void RB_CalcBulgeVertexes(deformStage_t *ds)
 	const float *st     = ( const float * ) tess.texCoords[0];
 	float       *xyz    = ( float * ) tess.xyz;
 	float       *normal = ( float * ) tess.normal;
-	float       now     = backEnd.refdef.time * ds->bulgeSpeed * 0.001f;
-	int         off;
+	double      now     = backEnd.refdef.time * ds->bulgeSpeed * 0.001;
+	int64_t     off;
 	float       scale;
 
 	for (i = 0; i < tess.numVertexes; i++, xyz += 4, st += 4, normal += 4)
 	{
-		off   = (float)(FUNCTABLE_SIZE / (M_PI * 2)) * (st[0] * ds->bulgeWidth + now);
+		off   = (FUNCTABLE_SIZE / (M_PI * 2)) * (st[0] * ds->bulgeWidth + now);
 		scale = tr.sinTable[off & FUNCTABLE_MASK] * ds->bulgeHeight;
 
 		xyz[0] += normal[0] * scale;
@@ -305,7 +305,7 @@ void RB_CalcMoveVertexes(deformStage_t *ds)
 	int    i;
 	float  *xyz;
 	float  *table;
-	double scale;
+	float  scale;
 	vec3_t offset;
 
 	table = TableForFunc(ds->deformationWave.func);
